+++ backend/canisters/community/CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 
 - Bot's can't access private channels they don't own ([#8181](https://github.com/open-chat-labs/open-chat/pull/8181))
+- Use MessagePack to serialize `init` args ([#8183](https://github.com/open-chat-labs/open-chat/pull/8183))
 
 ### Removed